+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0b] - 2025-05-05
+
+### Changed
+- Internal: getAirline() function reworked.
+- 
+
+### Fixed
+- Competitor monitoring can now correctly fetch schedule and display the info of schedule on dashboard. [#2](https://github.com/NEWLY2014/AirlineSim-Enhancement-Suite/issues/2)
+- Various Misspellings corrected. 
+
 ## [0.7.0a] - 2025-05-04
 
 ### Changed